test(i18n): seed English chunk synchronously so component tests see real strings
Frontend t() returned raw i18n KEYS on the first synchronous render because
translation chunks load via dynamic import() in an effect (post-render).
Every component test querying by translated text failed ("Unable to find
.../Bug report/i") — ~366 of the 816 frontend failures.

Fix: LanguageProvider lazy-inits currentTranslations from the module cache
(a genuine improvement — reuses an already-parsed chunk on remount, no-op on
a cold first paint in production), and the test setup primes the cache with
the statically-bundled English dict. Loaders/cache moved to a dedicated
translationLoader module so the provider file still exports only the
component (react-refresh lint).

Frontend failures: 816 -> 450. FeedbackDialog suite green (also updates the
dropzone prompt assertion to the new "Drag a file here" copy).
